==> 04_community_composition.Rmd <==
# Community composition

```{r load_data_community}
load("data/data.Rdata")
```

## Taxonomy overview 

**Mean percentage of Bacteria and Archaea in the samples**

```{r bateria_archaea_summary, warning=FALSE, comments="", message=FALSE}
domain_summary <- genome_counts_filt %>%
  mutate_at(vars(-genome),~./sum(.)) %>% #apply TSS normalisation
  pivot_longer(-genome, names_to = "sample", values_to = "count") %>%
  left_join(sample_metadata, by = join_by(sample == sample)) %>%
  left_join(genome_metadata, by = join_by(genome == genome)) %>%
  group_by(sample,domain) %>%
  summarise(relabun=sum(count))

domain_summary %>%
    group_by(domain) %>%
    summarise(mean=mean(relabun, na.rm=T)*100,sd=sd(relabun, na.rm=T)*100) %>%
    arrange(-mean) %>%
    tt()
```

```{r taxonomy_barplot, fig.height=6, fig.width=10, fig.fullwidth=TRUE}
#ggsave("figures/mtcars.png", width = 10, height = 6)
genome_counts_filt %>%
  mutate_at(vars(-genome),~./sum(.)) %>% #apply TSS normalisation
  pivot_longer(-genome, names_to = "sample", values_to = "count") %>% #reduce to minimum number of columns
  left_join(., genome_metadata, by = join_by(genome == genome)) %>% #append genome metadata
  left_join(., sample_metadata, by = join_by(sample == sample)) %>% #append sample metadata
  filter(count > 0) %>% #filter 0 counts
#  filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  mutate(season=factor(season,levels=c("spring","autumn"))) %>% 
  ggplot(., aes(x=sample,y=count, fill=phylum, group=phylum)) + #grouping enables keeping the same sorting of taxonomic units
    geom_bar(stat="identity", colour="white", linewidth=0.1, show.legend = FALSE) + #plot stacked bars with white borders
    scale_fill_manual(values=phylum_colors) +
    facet_nested(. ~ type + region+season,  scales="free") + #facet per day and treatment
    guides(fill = guide_legend(ncol = 2)) +
    theme(axis.text.x = element_blank(),
          axis.ticks.x = element_blank(),
          axis.title.x = element_blank(),
          panel.background = element_blank(),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          axis.line = element_line(linewidth = 0.5, linetype = "solid", colour = "black"),
          panel.border = element_rect(colour = "black", fill = NA),
          strip.background = element_rect(fill = "white", color = "black"),
          strip.text = element_text(size = 8, lineheight = 0.6)) +
   labs(fill="Phylum",y = "Relative abundance",x="Samples")
```

**Number of bacteria phyla**

```{r phyla, comment="", echo=FALSE, message=FALSE, warning=FALSE}
genome_metadata %>% 
  filter(domain == "d__Bacteria")%>%
  dplyr::select(phylum) %>%
  unique() %>%
  pull() %>%
  length()
```

**Bacteria phyla in natural ponds**

```{r phyla_nat, comment="", echo=FALSE, message=FALSE, warning=FALSE}
physeq_genome_sub <- subset_samples(physeq_genome, type == "natural")
physeq_sample_filt <- prune_taxa(taxa_sums(physeq_genome_sub)>0, physeq_genome_sub)
data.frame(physeq_sample_filt@tax_table) %>% 
  filter(domain == "d__Bacteria")%>% 
  dplyr::select(phylum) %>% 
  unique()
```

**Bacteria phyla in protected ponds**

```{r phyla_prot, comment="", echo=FALSE, message=FALSE, warning=FALSE}
physeq_genome_sub <- subset_samples(physeq_genome, type == "protected")
physeq_sample_filt <- prune_taxa(taxa_sums(physeq_genome_sub)>0, physeq_genome_sub)
data.frame(physeq_sample_filt@tax_table) %>% 
  filter(domain == "d__Bacteria")%>% 
  dplyr::select(phylum) %>% 
  unique()
```

**Bacteria phyla in urban ponds**

```{r phyla_urb, comment="", echo=FALSE, message=FALSE, warning=FALSE}
physeq_genome_sub <- subset_samples(physeq_genome, type == "urban")
physeq_sample_filt <- prune_taxa(taxa_sums(physeq_genome_sub)>0, physeq_genome_sub)
data.frame(physeq_sample_filt@tax_table) %>% 
  filter(domain == "d__Bacteria")%>% 
  dplyr::select(phylum) %>% 
  unique()
```

**Number of Archaea phyla**

```{r arch, comment="", echo=FALSE, message=FALSE, warning=FALSE}
genome_metadata %>% 
  filter(domain == "d__Archaea")%>%
  dplyr::select(phylum) %>%
  unique() %>%
  pull() %>%
  length()
```

**Archaea phyla in natural ponds**

```{r arch_nat, comment="", echo=FALSE, message=FALSE, warning=FALSE}
physeq_genome_sub <- subset_samples(physeq_genome, type == "natural")
physeq_sample_filt <- prune_taxa(taxa_sums(physeq_genome_sub)>0, physeq_genome_sub)
data.frame(physeq_sample_filt@tax_table) %>% 
  filter(domain == "d__Archaea")%>% 
  dplyr::select(phylum) %>% 
  unique()
```

**Archaea phyla in protected ponds**

```{r arch_prot, comment="", echo=FALSE, message=FALSE, warning=FALSE}
physeq_genome_sub <- subset_samples(physeq_genome, type == "protected")
physeq_sample_filt <- prune_taxa(taxa_sums(physeq_genome_sub)>0, physeq_genome_sub)
data.frame(physeq_sample_filt@tax_table) %>% 
  filter(domain == "d__Archaea")%>% 
  dplyr::select(phylum) %>% 
  unique()
```

**Archaea phyla in urban ponds**

```{r arch_urb, comment="", echo=FALSE, message=FALSE, warning=FALSE}
physeq_genome_sub <- subset_samples(physeq_genome, type == "urban")
physeq_sample_filt <- prune_taxa(taxa_sums(physeq_genome_sub)>0, physeq_genome_sub)
data.frame(physeq_sample_filt@tax_table) %>% 
  filter(domain == "d__Archaea")%>% 
  dplyr::select(phylum) %>% 
  unique()
```

## Phylum relative abundances
```{r taxonomy_phylum, warning=FALSE, comments="", message=FALSE}
phylum_summary <- genome_counts_filt %>%
  mutate_at(vars(-genome),~./sum(.)) %>% #apply TSS nornalisation
  pivot_longer(-genome, names_to = "sample", values_to = "count") %>%
  left_join(sample_metadata, by = join_by(sample == sample)) %>%
  left_join(genome_metadata, by = join_by(genome == genome)) %>%
  group_by(sample,phylum,type, season) %>%
  summarise(relabun=sum(count))
```
```{r taxonomy_phylum_summary_total, warning=FALSE, comments="", message=FALSE}
phylum_summary %>%
    group_by(phylum) %>%
    summarise(total_mean=mean(relabun*100, na.rm=T),
              total_sd=sd(relabun*100, na.rm=T))  %>%
    mutate(total=str_c(round(total_mean,2),"±",round(total_sd,2))) %>% 
    arrange(-total_mean) %>% 
    dplyr::select(phylum,total) %>% 
    tt()
```
```{r taxonomy_boxplot_phylum, warning=FALSE, comments="", message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
phylum_arrange <- phylum_summary %>%
    group_by(phylum) %>%
    summarise(mean=mean(relabun)) %>%
    arrange(-mean) %>%
    dplyr::select(phylum) %>%
    pull()

phylum_summary %>%
    filter(phylum %in% phylum_arrange) %>%
    mutate(phylum=factor(phylum,levels=rev(phylum_arrange))) %>%
    ggplot(aes(x=relabun, y=phylum, group=phylum, color=phylum)) +
        scale_color_manual(values=phylum_colors[rev(phylum_arrange)]) +
        geom_jitter(alpha=0.5, show.legend = FALSE) + 
        theme_minimal() + 
        theme(legend.position="none") +
        labs(y="Phylum",x="Relative abundance")
```

### Per pond
```{r taxonomy_phylum_summary_type, warning=FALSE, comments="", message=FALSE}
phylum_summary %>%
    group_by(phylum) %>%
    summarise(total_mean=mean(relabun*100, na.rm=T),
              total_sd=sd(relabun*100, na.rm=T),
              natural_mean=mean(relabun[type=="natural"]*100, na.rm=T),
              natural_sd=sd(relabun[type=="natural"]*100, na.rm=T),
              protected_mean=mean(relabun[type=="protected"]*100, na.rm=T),
              protected_sd=sd(relabun[type=="protected"]*100, na.rm=T),
              urban_mean=mean(relabun[type=="urban"]*100, na.rm=T),
              urban_sd=sd(relabun[type=="urban"]*100, na.rm=T))  %>%
    mutate(total=str_c(round(total_mean,2),"±",round(total_sd,2)),
           natural=str_c(round(natural_mean,2),"±",round(natural_sd,2)),
           protected=str_c(round(protected_mean,2),"±",round(protected_sd,2)),
           urban=str_c(round(urban_mean,2),"±",round(urban_sd,2))) %>% 
    arrange(-total_mean) %>% 
    dplyr::select(phylum,total,natural,protected,urban) %>% 
    tt()
```

```{r taxonomy_boxplot_phylum_type, warning=FALSE, comments="", message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
phylum_summary %>%
  filter(phylum %in% phylum_arrange) %>%
  mutate(phylum=factor(phylum,levels=rev(phylum_arrange))) %>%
  filter(relabun > 0) %>%
  ggplot(aes(x=relabun, y=phylum, group=phylum, color=phylum)) +
  scale_color_manual(values=phylum_colors[-8]) +
  geom_jitter(alpha=0.5, show.legend = FALSE) + 
  facet_grid(.~type)+
  theme_minimal() + 
  labs(y="Family", x="Relative abundance", color="Phylum")
```

### Per season
```{r taxonomy_phylum_summary_season, warning=FALSE, comments="", message=FALSE}
phylum_summary %>%
    group_by(phylum) %>% 
    summarise(total_mean=mean(relabun*100, na.rm=T),
              total_sd=sd(relabun*100, na.rm=T),
              spring_mean=mean(relabun[season=="spring"]*100, na.rm=T),
              spring_sd=sd(relabun[season=="spring"]*100, na.rm=T),
              autumn_mean=mean(relabun[season=="autumn"]*100, na.rm=T),
              autumn_sd=sd(relabun[season=="autumn"]*100, na.rm=T))  %>%
    mutate(total=str_c(round(total_mean,2),"±",round(total_sd,2)),
           spring=str_c(round(spring_mean,2),"±",round(spring_sd,2)),
           autumn=str_c(round(autumn_mean,2),"±",round(autumn_sd,2))) %>% 
    arrange(-total_mean) %>% 
    dplyr::select(phylum,total,spring,autumn) %>% 
    tt()
```

```{r taxonomy_boxplot_phylum_season, warning=FALSE, comments="", message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
phylum_summary %>%
  filter(phylum %in% phylum_arrange) %>%
  mutate(phylum=factor(phylum,levels=rev(phylum_arrange))) %>%
  filter(relabun > 0) %>%
  ggplot(aes(x=relabun, y=phylum, group=phylum, color=phylum)) +
  scale_color_manual(values=phylum_colors[-8]) +
  geom_jitter(alpha=0.5, show.legend = FALSE) + 
  facet_grid(.~season)+
  theme_minimal() + 
  labs(y="Family", x="Relative abundance", color="Phylum")
```

## Family relative abundances
```{r taxonomy_family_summary, warning=FALSE, comments="", message=FALSE}
family_summary <- genome_counts_filt %>%
  mutate_at(vars(-genome),~./sum(.)) %>% #apply TSS nornalisation
  pivot_longer(-genome, names_to = "sample", values_to = "count") %>% #reduce to minimum number of columns
  left_join(sample_metadata, by = join_by(sample == sample)) %>% #append sample metadata
  left_join(., genome_metadata, by = join_by(genome == genome)) %>% #append genome metadata
  group_by(sample,family, type, region, season) %>%
  summarise(relabun=sum(count))

family_arrange <- family_summary %>%
    group_by(family) %>%
    summarise(mean=sum(relabun)) %>%
    arrange(-mean) %>%
    select(family) %>%
    pull()

family_summary %>%
    left_join(genome_metadata %>% dplyr::select(family,phylum) %>% unique(),by=join_by(family==family)) %>%
#    left_join(sample_metadata,by=join_by(sample==sample)) %>%
    filter(family %in% family_arrange[1:20]) %>%
    mutate(family=factor(family,levels=rev(family_arrange[1:20]))) %>%
    filter(relabun > 0) %>%
    ggplot(aes(x=relabun, y=family, group=family, color=phylum)) +
        scale_color_manual(values=phylum_colors) +
        geom_jitter(alpha=0.5) + 
        theme_minimal() + 
        labs(y="Family", x="Relative abundance", color="Phylum")
```

### Per pond
```{r taxonomy_family_summary_type, warning=FALSE, comments="", message=FALSE}
family_summary %>%
    group_by(family) %>%
    summarise(total_mean=mean(relabun*100, na.rm=T),
              total_sd=sd(relabun*100, na.rm=T),
              natural_mean=mean(relabun[type=="natural"]*100, na.rm=T),
              natural_sd=sd(relabun[type=="natural"]*100, na.rm=T),
              protected_mean=mean(relabun[type=="protected"]*100, na.rm=T),
              protected_sd=sd(relabun[type=="protected"]*100, na.rm=T),
              urban_mean=mean(relabun[type=="urban"]*100, na.rm=T),
              urban_sd=sd(relabun[type=="urban"]*100, na.rm=T))  %>%
    mutate(total=str_c(round(total_mean,2),"±",round(total_sd,2)),
           natural=str_c(round(natural_mean,2),"±",round(natural_sd,2)),
           protected=str_c(round(protected_mean,2),"±",round(protected_sd,2)),
           urban=str_c(round(urban_mean,2),"±",round(urban_sd,2))) %>% 
    arrange(-total_mean) %>% 
    dplyr::select(family,total,natural,protected,urban) %>% 
    tt()
```

```{r taxonomy_jitterplot_family_type, warning=FALSE, comments="", message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
family_arrange <- family_summary %>%
    group_by(family) %>%
    summarise(mean=sum(relabun)) %>%
    arrange(-mean) %>%
    dplyr::select(family) %>%
    pull()

family_summary %>%
    left_join(genome_metadata %>% dplyr::select(family,phylum) %>% unique(),by=join_by(family==family)) %>%
#    left_join(sample_metadata,by=join_by(sample==sample)) %>%
    filter(family %in% family_arrange[1:20]) %>%
    mutate(family=factor(family,levels=rev(family_arrange[1:20]))) %>%
    filter(relabun > 0) %>%
    ggplot(aes(x=relabun, y=family, group=family, color=phylum)) +
        scale_color_manual(values=phylum_colors) +
        geom_jitter(alpha=0.5) + 
        facet_grid(.~type)+
        theme_minimal() + 
        labs(y="Family", x="Relative abundance", color="Phylum")

# Per origin
# family_summary %>%
#     left_join(genome_metadata %>% dplyr::select(family,phylum) %>% unique(),by=join_by(family==family)) %>%
# #    left_join(sample_metadata,by=join_by(sample==sample)) %>%
#     filter(family %in% family_arrange[1:20]) %>%
#     mutate(family=factor(family,levels=rev(family_arrange[1:20]))) %>%
#     filter(relabun > 0) %>%
#     ggplot(aes(x=relabun, y=family, group=family, color=phylum)) +
#         scale_color_manual(values=phylum_colors[-8]) +
#         geom_jitter(alpha=0.5) + 
#         facet_grid(.~region)+
#         theme_minimal() + 
#         labs(y="Family", x="Relative abundance", color="Phylum")


```

### Per season
```{r taxonomy_family_summary_season, warning=FALSE, comments="", message=FALSE}
family_summary %>%
    group_by(family) %>% 
    summarise(total_mean=mean(relabun*100, na.rm=T),
              total_sd=sd(relabun*100, na.rm=T),
              spring_mean=mean(relabun[season=="spring"]*100, na.rm=T),
              spring_sd=sd(relabun[season=="spring"]*100, na.rm=T),
              autumn_mean=mean(relabun[season=="autumn"]*100, na.rm=T),
              autumn_sd=sd(relabun[season=="autumn"]*100, na.rm=T))  %>%
    mutate(total=str_c(round(total_mean,2),"±",round(total_sd,2)),
           spring=str_c(round(spring_mean,2),"±",round(spring_sd,2)),
           autumn=str_c(round(autumn_mean,2),"±",round(autumn_sd,2))) %>% 
    arrange(-total_mean) %>% 
    dplyr::select(family,total,spring,autumn) %>% 
    tt()
```

```{r taxonomy_boxplot_family_season, warning=FALSE, comments="", message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
family_summary %>%
    left_join(genome_metadata %>% dplyr::select(family,phylum) %>% unique(),by=join_by(family==family)) %>%
#    left_join(sample_metadata,by=join_by(sample==sample)) %>%
    filter(family %in% family_arrange[1:20]) %>%
    mutate(family=factor(family,levels=rev(family_arrange[1:20]))) %>%
    filter(relabun > 0) %>%
    ggplot(aes(x=relabun, y=family, group=family, color=phylum)) +
        scale_color_manual(values=phylum_colors) +#[-8]
        geom_jitter(alpha=0.5) + 
        facet_grid(.~season)+
        theme_minimal() + 
        labs(y="Family", x="Relative abundance", color="Phylum")
```

## Genus relative abundances
```{r taxonomy_genus, warning=FALSE, comments="", message=FALSE}
genus_summary <- genome_counts_filt %>%
  mutate_at(vars(-genome),~./sum(.)) %>% #apply TSS nornalisation
  pivot_longer(-genome, names_to = "sample", values_to = "count") %>% #reduce to minimum number of columns
  left_join(sample_metadata, by = join_by(sample == sample)) %>% #append sample metadata
  left_join(genome_metadata, by = join_by(genome == genome)) %>% #append genome metadata
  group_by(sample,phylum,genus, type, season) %>%
  summarise(relabun=sum(count)) %>%
  filter(genus != "g__") %>%
  mutate(genus= sub("^g__", "", genus))
```

### Per pond
```{r taxonomy_genus_summary_type, warning=FALSE, comments="", message=FALSE}
genus_summary %>%
    group_by(genus) %>%
    summarise(total_mean=mean(relabun*100, na.rm=T),
              total_sd=sd(relabun*100, na.rm=T),
              natural_mean=mean(relabun[type=="natural"]*100, na.rm=T),
              natural_sd=sd(relabun[type=="natural"]*100, na.rm=T),
              protected_mean=mean(relabun[type=="protected"]*100, na.rm=T),
              protected_sd=sd(relabun[type=="protected"]*100, na.rm=T),
              urban_mean=mean(relabun[type=="urban"]*100, na.rm=T),
              urban_sd=sd(relabun[type=="urban"]*100, na.rm=T))  %>%
    mutate(total=str_c(round(total_mean,2),"±",round(total_sd,2)),
           natural=str_c(round(natural_mean,2),"±",round(natural_sd,2)),
           protected=str_c(round(protected_mean,2),"±",round(protected_sd,2)),
           urban=str_c(round(urban_mean,2),"±",round(urban_sd,2))) %>% 
    arrange(-total_mean) %>% 
    dplyr::select(genus,total,natural,protected,urban) %>% 
    tt()
```

```{r taxonomy_jitterplot_genus_type, fig.height=14, fig.width=10, fig.fullwidth=TRUE}
genus_summary %>%
  #    left_join(sample_metadata,by=join_by(sample==sample)) %>%
  mutate(genus=factor(genus, levels=rev(genus_summary_sort %>% pull(genus)))) %>%
  filter(relabun > 0) %>%
  ggplot(aes(x=relabun, y=genus, group=genus, color=phylum)) +
  scale_color_manual(values=phylum_colors) +
  geom_jitter(alpha=0.5) + 
  facet_grid(.~type)+
  theme_minimal() +
  theme(
    axis.text = element_text(size=6),
    )+
        labs(y="Genus", x="Relative abundance", color="Phylum")

```

### Per season
```{r taxonomy_genus_summary_season, warning=FALSE, comments="", message=FALSE}
genus_summary %>%
    group_by(genus) %>%
    summarise(total_mean=mean(relabun*100, na.rm=T),
              total_sd=sd(relabun*100, na.rm=T),
              spring_mean=mean(relabun[season=="spring"]*100, na.rm=T),
              spring_sd=sd(relabun[season=="spring"]*100, na.rm=T),
              autumn_mean=mean(relabun[season=="autumn"]*100, na.rm=T),
              autumn_sd=sd(relabun[season=="autumn"]*100, na.rm=T))  %>%
    mutate(total=str_c(round(total_mean,2),"±",round(total_sd,2)),
           spring=str_c(round(spring_mean,2),"±",round(spring_sd,2)),
           autumn=str_c(round(autumn_mean,2),"±",round(autumn_sd,2))) %>% 
    arrange(-total_mean) %>% 
    dplyr::select(genus,total,spring,autumn) %>% 
    tt()

genus_summary_sort <- genus_summary %>%
    group_by(genus) %>%
    summarise(mean=mean(relabun, na.rm=T),sd=sd(relabun, na.rm=T)) %>%
    arrange(-mean) 

```

```{r taxonomy_jitterplot_genus_season, fig.height=14, fig.width=10, fig.fullwidth=TRUE}
genus_summary %>%
  #    left_join(sample_metadata,by=join_by(sample==sample)) %>%
  mutate(genus=factor(genus, levels=rev(genus_summary_sort %>% pull(genus)))) %>%
  filter(relabun > 0) %>%
  ggplot(aes(x=relabun, y=genus, group=genus, color=phylum)) +
  scale_color_manual(values=phylum_colors) +
  geom_jitter(alpha=0.5) + 
  facet_grid(.~season)+
  theme_minimal() +
  theme(
    axis.text = element_text(size=6),
    )+
        labs(y="Genus", x="Relative abundance", color="Phylum")

```


## Archaea
```{r archaea, comment="", echo=FALSE, message=FALSE, warning=FALSE}
physeq_archaea <- subset_taxa(physeq_genome, domain == "d__Archaea")
cat(nrow(physeq_archaea@tax_table))
unique(physeq_archaea@tax_table) %>% 
  as.data.frame() %>%
  paged_table() 
```

```{r barplot_archaea, comment="", echo=FALSE, message=FALSE, warning=FALSE}
metadata_row_archaea <- data.frame(sample_data(physeq_archaea)) %>% rownames_to_column(., "sample")
asv_table_archaea <- data.frame(physeq_archaea@otu_table)%>% rownames_to_column(., "genome")
taxonomy_table_archaea <- data.frame(physeq_archaea@tax_table)%>% rownames_to_column(., "genome")

asv_table_archaea %>%
  mutate_at(vars(-genome),~./sum(.)) %>% #apply TSS normalisation
  pivot_longer(-genome, names_to = "sample", values_to = "count") %>% #reduce to minimum number of columns
  left_join(., taxonomy_table_archaea, by = join_by(genome == genome)) %>% #append genome metadata
  left_join(., metadata_row_archaea, by = join_by(sample == sample)) %>% #append sample metadata
  ggplot(., aes(x=sample,y=count, fill=phylum, group=phylum)) + #grouping enables keeping the same sorting of taxonomic units
    geom_bar(stat="identity", colour="white", linewidth=0.1) +
#    scale_fill_manual(values=phylum_colors) +
  facet_grid(.~type,  scales="free_x")+
 #   facet_nested(.~type+region,  scales="free_x") +
theme(axis.text.x = element_blank(),
          axis.ticks.x = element_blank(),
          axis.title.x = element_blank(),
          panel.background = element_blank(),
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          axis.line = element_line(linewidth = 0.5, linetype = "solid", colour = "black"),
 #         panel.border = element_rect(colour = "black", fill = NA),
          strip.background = element_rect(fill = "white", color = "black"),
          strip.text = element_text(size = 8, lineheight = 0.6)
      ) +
  labs(x = "Sample", y = "Relative abundances",fill="Phylum")+
    guides(fill = guide_legend(ncol = 1)) 
```

### Diversity
```{r archaea_div, comment="", echo=FALSE, message=FALSE, warning=FALSE}
richness <- data.frame(physeq_archaea@otu_table) %>%
#  column_to_rownames(var = "genome") %>%
  dplyr::select(where(~ !all(. == 0))) %>%
  hilldiv(., q = 0) %>%
  t() %>%
  as.data.frame() %>%
  dplyr::rename(richness = 1) %>%
  rownames_to_column(var = "sample")
neutral <- data.frame(physeq_archaea@otu_table) %>%
#  column_to_rownames(var = "genome") %>%
  dplyr::select(where(~ !all(. == 0))) %>%
  hilldiv(., q = 1) %>%
  t() %>%
  as.data.frame() %>%
  dplyr::rename(neutral = 1) %>%
  rownames_to_column(var = "sample")

metadata_archaea <- data.frame(tax_table(physeq_archaea))%>% rownames_to_column(., "genome")
archaea_tree <- keep.tip(genome_tree, tip=metadata_archaea$genome) 
phylogenetic <- data.frame(physeq_archaea@otu_table) %>%
#  column_to_rownames(var = "genome") %>%
  dplyr::select(where(~ !all(. == 0))) %>%
  hilldiv(., q = 1, tree = archaea_tree) %>%
  t() %>%
  as.data.frame() %>%
  dplyr::rename(phylogenetic = 1) %>%
  rownames_to_column(var = "sample")

# Aggregate basal GIFT into elements
counts_archaea <- data.frame(physeq_archaea@otu_table) %>% 
  as.data.frame() %>% 
  rownames_to_column(., "genome")
counts_archaea_func <- counts_archaea[counts_archaea$genome %in% rownames(genome_gifts),]
genome_gifts_archaea <- genome_gifts[counts_archaea$genome,]

dist <- genome_gifts_archaea %>%
  to.elements(., GIFT_db) %>%
  traits2dist(., method = "gower")

functional <- counts_archaea %>%
  filter(genome %in% rownames(dist)) %>%
  column_to_rownames(var = "genome") %>%
#  dplyr::select(where(~ !all(. == 0))) %>%
  hilldiv(., q = 1, dist = dist) %>%
  t() %>%
  as.data.frame() %>%
  dplyr::rename(functional = 1) %>%
  rownames_to_column(var = "sample") %>%
  mutate(functional = if_else(is.nan(functional), 1, functional))

# Merge all metrics
alpha_div <- richness %>%
  full_join(neutral, by = join_by(sample == sample)) %>%
  full_join(phylogenetic, by = join_by(sample == sample)) %>%
  full_join(functional, by = join_by(sample == sample))

alpha_div
```

#### Richness
```{r archaea_div_rich, comment="", echo=FALSE, message=FALSE, warning=FALSE}
alpha_div %>%
  pivot_longer(-sample, names_to = "metric", values_to = "value") %>%
  left_join(., sample_metadata, by = join_by(sample == sample)) %>%
  filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  filter(metric=="richness") %>%
  ggplot(aes(y = value, x = region, group=region, color=region, fill=region)) +
  geom_jitter()+
  geom_boxplot(alpha=0.5, outlier.shape = NA)  +
  scale_color_manual(values=location_colors)+
  scale_fill_manual(values=location_colors) +
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.text.x = element_text(angle = 45, hjust = 1)
      )
```


#### Neutral
```{r archaea_div_neutral, comment="", echo=FALSE, message=FALSE, warning=FALSE}
alpha_div %>%
  pivot_longer(-sample, names_to = "metric", values_to = "value") %>%
  left_join(., sample_metadata, by = join_by(sample == sample)) %>%
  filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  filter(metric=="neutral") %>%
  ggplot(aes(y = value, x = region, group=region, color=region, fill=region)) +
  geom_jitter()+
  geom_boxplot(alpha=0.5, outlier.shape = NA)  +
  scale_color_manual(values=location_colors)+
  scale_fill_manual(values=location_colors) +
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.text.x = element_text(angle = 45, hjust = 1)
      )
```


#### Phylogenetic
```{r archaea_div_phylo, comment="", echo=FALSE, message=FALSE, warning=FALSE}
alpha_div %>%
  pivot_longer(-sample, names_to = "metric", values_to = "value") %>%
  left_join(., sample_metadata, by = join_by(sample == sample)) %>%
  filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  filter(metric=="phylogenetic") %>%
  ggplot(aes(y = value, x = region, group=region, color=region, fill=region)) +
  geom_jitter()+
  geom_boxplot(alpha=0.5, outlier.shape = NA)  +
  scale_color_manual(values=location_colors)+
  scale_fill_manual(values=location_colors) +
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.text.x = element_text(angle = 45, hjust = 1)
      )
```
