==> 05_alpha_diversity.Rmd <==
# Alpha diversity

```{r load_data_alpha, comment="", message=FALSE, warning=FALSE, eval=FALSE}
load("data/data.Rdata")
```
```{r load_data_mag_filtdamr1, comment="", message=FALSE, warning=FALSE}
load("data/data_filtdamr_80.Rdata")
sample_metadata <- sample_metadata_filtdamr
genome_counts_filt <- genome_counts_filtdamr
genome_metadata <- genome_metadata_filtdamr
genome_tree <- genome_tree_filtdamr
genome_gifts <- genome_gifts_filtdamr
phylum_colors <- phylum_colors_filtdamr
```

```{r alpha_div, comment="", message=FALSE, warning=FALSE}
# Calculate Hill numbers
richness <- genome_counts_filt %>%
  column_to_rownames(var = "genome") %>%
  dplyr::select(where(~ !all(. == 0))) %>%
  hilldiv(., q = 0) %>%
  t() %>%
  as.data.frame() %>%
  dplyr::rename(richness = 1) %>%
  rownames_to_column(var = "sample")

neutral <- genome_counts_filt %>%
  column_to_rownames(var = "genome") %>%
  dplyr::select(where(~ !all(. == 0))) %>%
  hilldiv(., q = 1) %>%
  t() %>%
  as.data.frame() %>%
  dplyr::rename(neutral = 1) %>%
  rownames_to_column(var = "sample")

phylogenetic <- genome_counts_filt %>%
  column_to_rownames(var = "genome") %>%
  dplyr::select(where(~ !all(. == 0))) %>%
  hilldiv(., q = 1, tree = genome_tree) %>%
  t() %>%
  as.data.frame() %>%
  dplyr::rename(phylogenetic = 1) %>%
  rownames_to_column(var = "sample")

# Aggregate basal GIFT into elements
dist <- genome_gifts %>%
  to.elements(., GIFT_db) %>%
  traits2dist(., method = "gower")

functional <- genome_counts_filt %>%
  filter(genome %in% rownames(dist)) %>% 
  column_to_rownames(var = "genome") %>%
  dplyr::select(where(~ !all(. == 0))) %>%
  hilldiv(., q = 1, dist = dist) %>%
  t() %>%
  as.data.frame() %>%
  dplyr::rename(functional = 1) %>%
  rownames_to_column(var = "sample") %>%
  mutate(functional = if_else(is.nan(functional), 1, functional))

# Merge all metrics
alpha_div <- richness %>%
  full_join(neutral, by = join_by(sample == sample)) %>%
  full_join(phylogenetic, by = join_by(sample == sample)) %>%
  full_join(functional, by = join_by(sample == sample))


# sample_metadata_div <- sample_metadata %>% 
#   left_join(alpha_div, by = join_by(sample == sample))
# 
# write_csv(sample_metadata_div, "data/sample_metadata_div.csv")  
```

## Types of pond
```{r alpha_div_types_summary, comment="",echo=FALSE, message=FALSE, warning=FALSE}
alpha_div %>%
  pivot_longer(-sample, names_to = "alpha", values_to = "value") %>%
  left_join(sample_metadata, by = join_by(sample == sample)) %>%
    group_by(alpha)%>%
    summarise(
              natural_mean=mean(value[type=="natural"], na.rm=T),
              natural_sd=sd(value[type=="natural"], na.rm=T),
              protected_mean=mean(value[type=="protected"], na.rm=T),
              protected_sd=sd(value[type=="protected"], na.rm=T),
              urban_mean=mean(value[type=="urban"], na.rm=T),
              urban_sd=sd(value[type=="urban"], na.rm=T)) %>%
    mutate(
           natural=str_c(round(natural_mean,2),"±",round(natural_sd,2)),
           protected=str_c(round(protected_mean,2),"±",round(protected_sd,2)),
           urban=str_c(round(urban_mean,2),"±",round(urban_sd,2))) %>% 
    arrange(-natural_mean) %>% 
    dplyr::select(alpha,natural,protected,urban) %>% 
    tt()
```

```{r alpha_div_boxplot, comment="",echo=FALSE, message=FALSE, warning=FALSE}
#Richness
plot1 <- alpha_div %>%
  pivot_longer(-sample, names_to = "metric", values_to = "value") %>%
  left_join(., sample_metadata, by = join_by(sample == sample)) %>%
#  filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  filter(metric=="richness") %>%
      ggplot(aes(y = value, x = type, group=type, color=type, fill=type, shape=season)) +
      geom_jitter(width = 0.2, show.legend = FALSE) +
      geom_boxplot(width = 0.5, alpha=0.5,outlier.shape = NA, show.legend = FALSE) +
  scale_color_manual(values=type_colors)+
  scale_fill_manual(values=type_colors) +
  scale_x_discrete(labels=c("natural" = "Natural", "protected" = "Protected",
                            "urban" = "Urban")) +
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.text = element_text(size=10),
    axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)),
    axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0))
    )+
  labs(x = "Type of pond", y = "Richness")

#Neutral
plot2 <- alpha_div %>%
  pivot_longer(-sample, names_to = "metric", values_to = "value") %>%
  left_join(., sample_metadata, by = join_by(sample == sample)) %>%
#  filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  filter(metric=="neutral") %>%
  ggplot(aes(y = value, x = type, group=type, color=type, fill=type)) +
  geom_jitter(width = 0.2, show.legend = FALSE) +
  geom_boxplot(width = 0.5, alpha=0.5,outlier.shape = NA, show.legend = FALSE) +
  scale_color_manual(values=type_colors)+
  scale_fill_manual(values=type_colors) +
  scale_x_discrete(labels=c("natural" = "Natural", "protected" = "Protected",
                            "urban" = "Urban")) +
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.text = element_text(size=10),
    axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)),
    axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0))
    )+
  labs(x = "Type of pond", y = "Neutral")

#Phylogenetic
plot3 <- alpha_div %>%
  pivot_longer(-sample, names_to = "metric", values_to = "value") %>%
  left_join(., sample_metadata, by = join_by(sample == sample)) %>%
 # filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  filter(metric=="phylogenetic") %>%
  ggplot(aes(y = value, x = type, group=type, color=type, fill=type)) +
  geom_jitter(width = 0.2, show.legend = FALSE) +
  geom_boxplot(width = 0.5, alpha=0.5,outlier.shape = NA, show.legend = FALSE) +
  scale_color_manual(values=type_colors)+
  scale_fill_manual(values=type_colors) +
  scale_x_discrete(labels=c("natural" = "Natural", "protected" = "Protected",
                            "urban" = "Urban")) +
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.text = element_text(size=10),
    axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)),
    axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0))
    )+
  labs(x = "Type of pond", y = "Phylogenetic")

#Functional
plot4 <- alpha_div %>%
  pivot_longer(-sample, names_to = "metric", values_to = "value") %>%
  left_join(., sample_metadata, by = join_by(sample == sample)) %>%
  filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  filter(metric=="functional") %>%
  ggplot(aes(y = value, x = type, group=type, color=type, fill=type)) +
  geom_jitter(width = 0.2, show.legend = FALSE) +
  geom_boxplot(width = 0.5, alpha=0.5,outlier.shape = NA, show.legend = FALSE) +
  scale_color_manual(values=type_colors)+
  scale_fill_manual(values=type_colors) +
  scale_x_discrete(labels=c("natural" = "Natural", "protected" = "Protected",
                            "urban" = "Urban")) +
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.text = element_text(size=10),
    axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)),
    axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0))
    )+
  labs(x = "Type of pond", y = "Functional")
```
```{r div_plot_together, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
grid.arrange(arrangeGrob(plot1,plot2,plot3, plot4, ncol = 2))
```

### Mixed models

```{r mixed_rich_type, comment="", echo=FALSE, message=FALSE, warning=FALSE}
alpha_div_meta <- alpha_div %>%
  left_join(sample_metadata, by = join_by(sample == sample))

Model_richness <- glmer.nb(richness ~ type+(1|region), data = alpha_div_meta)
summary(Model_richness)
emmeans(Modelq0GLMMNB, pairwise ~ type)
```

```{r mixed_neutral_type, comment="", echo=FALSE, message=FALSE, warning=FALSE}

Model_neutral <- lme(fixed = neutral ~ type, data = alpha_div_meta,
               random = ~ 1 | region)#log(seq_depth)+
summary(Model_neutral)
emmeans(Model_neutral, pairwise ~ type)
```

```{r mixed_phylo_type, comment="", echo=FALSE, message=FALSE, warning=FALSE}
Model_phylo <- lme(fixed = phylogenetic ~ type, data = alpha_div_meta,
               random = ~ 1 | region)
summary(Model_phylo)
emmeans(Model_phylo, pairwise ~ type)
```

```{r mixed_func_type, comment="", echo=FALSE, message=FALSE, warning=FALSE}
Model_funct <- lme(fixed = functional ~ type, data = alpha_div_meta,
               random = ~ 1 | region)
summary(Model_funct)
emmeans(Model_funct, pairwise ~ type)
```

## Season
```{r alpha_div_season, comment="",echo=FALSE, message=FALSE, warning=FALSE}
season_colors <- c("#009270","#ac5c4c")
alpha_div %>%
  pivot_longer(-sample, names_to = "alpha", values_to = "value") %>%
  left_join(sample_metadata, by = join_by(sample == sample)) %>%
    group_by(alpha)%>%
    summarise(
              spring_mean=mean(value[season=="spring"], na.rm=T),
              spring_sd=sd(value[season=="spring"], na.rm=T),
              autumn_mean=mean(value[season=="autumn"], na.rm=T),
              autumn_sd=sd(value[season=="autumn"], na.rm=T)) %>%
    mutate(
           spring=str_c(round(spring_mean,2),"±",round(spring_sd,2)),
           autumn=str_c(round(autumn_mean,2),"±",round(autumn_sd,2))) %>% 
    arrange(-spring_mean) %>% 
    dplyr::select(alpha,spring,autumn) %>% 
    tt()
```


```{r alpha_div_boxplot_season, comment="",echo=FALSE, message=FALSE, warning=FALSE}
sample_metadata$season <- factor(sample_metadata$season, levels=c("spring", "autumn"))
#Richness
plot21 <- alpha_div %>%
  pivot_longer(-sample, names_to = "metric", values_to = "value") %>%
  left_join(., sample_metadata, by = join_by(sample == sample)) %>%
  filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  filter(metric=="richness") %>%
  ggplot(aes(y = value, x = season, group=season, color=season, fill=season)) +
  geom_jitter(width = 0.2, show.legend = FALSE) +
  geom_boxplot(width = 0.5, alpha=0.5,outlier.shape = NA, show.legend = FALSE) +
  scale_color_manual(values=season_colors)+
  scale_fill_manual(values=season_colors) +
  scale_x_discrete(labels=c("spring" = "Spring", "autumn" = "autumn")) +
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.text = element_text(size=10),
    axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)),
    axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0))
    )+
  labs(x = "Season", y = "Richness")

#Neutral
plot22 <- alpha_div %>%
  pivot_longer(-sample, names_to = "metric", values_to = "value") %>%
  left_join(., sample_metadata, by = join_by(sample == sample)) %>%
  filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  filter(metric=="neutral") %>%
  ggplot(aes(y = value, x = season, group=season, color=season, fill=season)) +
  geom_jitter(width = 0.2, show.legend = FALSE) +
  geom_boxplot(width = 0.5, alpha=0.5,outlier.shape = NA, show.legend = FALSE) +
  scale_color_manual(values=season_colors)+
  scale_fill_manual(values=season_colors) +
  scale_x_discrete(labels=c("spring" = "Spring", "autumn" = "autumn")) +
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.text = element_text(size=10),
    axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)),
    axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0))
    )+
  labs(x = "Season", y = "Neutral")

#Phylogenetic
plot23 <- alpha_div %>%
  pivot_longer(-sample, names_to = "metric", values_to = "value") %>%
  left_join(., sample_metadata, by = join_by(sample == sample)) %>%
  filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  filter(metric=="phylogenetic") %>%
  ggplot(aes(y = value, x = season, group=season, color=season, fill=season)) +
  geom_jitter(width = 0.2, show.legend = FALSE) +
  geom_boxplot(width = 0.5, alpha=0.5,outlier.shape = NA, show.legend = FALSE) +
  scale_color_manual(values=season_colors)+
  scale_fill_manual(values=season_colors) +
  scale_x_discrete(labels=c("spring" = "Spring", "autumn" = "autumn")) +
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.text = element_text(size=10),
    axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)),
    axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0))
    )+
  labs(x = "Season", y = "Phylogenetic")

#Functional
plot24 <- alpha_div %>%
  pivot_longer(-sample, names_to = "metric", values_to = "value") %>%
  left_join(., sample_metadata, by = join_by(sample == sample)) %>%
  filter(!region %in% c("Eskoriatza","Villabona")) %>% 
  filter(metric=="functional") %>%
  ggplot(aes(y = value, x = season, group=season, color=season, fill=season)) +
  geom_jitter(width = 0.2, show.legend = FALSE) +
  geom_boxplot(width = 0.5, alpha=0.5,outlier.shape = NA, show.legend = FALSE) +
  scale_color_manual(values=season_colors)+
  scale_fill_manual(values=season_colors) +
  scale_x_discrete(labels=c("spring" = "Spring", "autumn" = "autumn")) +
  coord_cartesian(xlim = c(1, NA)) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    panel.grid.minor.x = element_line(size = .1, color = "grey"),
    axis.text = element_text(size=10),
    axis.title.y = element_text(margin = margin(t = 0, r = 20, b = 0, l = 0)),
    axis.title.x = element_text(margin = margin(t = 20, r = 0, b = 0, l = 0))
    )+
  labs(x = "Season", y = "Functional")


```

```{r div_plot_together_season, comment="", echo=FALSE, message=FALSE, warning=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
grid.arrange(arrangeGrob(plot21,plot22,plot23, plot24, ncol = 2))
```

### Mixed models

```{r mixed_rich_season, comment="", echo=FALSE, message=FALSE, warning=FALSE}
alpha_div_meta <- alpha_div %>%
  left_join(sample_metadata, by = join_by(sample == sample))%>%
  filter(!region %in% c("Eskoriatza","Villabona"))

Model_richness <- glmer.nb(richness ~ season+(1|region), data = alpha_div_meta)
summary(Model_richness)

Model_neutral <- lme(fixed = neutral ~ season, data = alpha_div_meta,
               random = ~ 1 | region)#log(seq_depth)+
summary(Model_neutral)

Model_phylo <- lme(fixed = phylogenetic ~ season, data = alpha_div_meta,
               random = ~ 1 | region)
summary(Model_phylo)

Model_funct <- lme(fixed = functional ~ season, data = alpha_div_meta,
               random = ~ 1 | region)
summary(Model_funct)
```
